Fixed authorization and added resource links to creator user (#2)
* Very basic forum functionality works and is rendered mostly correctly

* Whn logging in I now pass the userId instead of the password. Because this is all behind a json api I cant see a reason why I need to access the username directly like this.

* Implemented UserResource
Added creator to Thread entity
Added link to creator from thread resource

* Exposed auth header in cors config
Abstracted out as much of the functionality for fetching the logfged in user and attaching to the creator field of any 'entity' upon creation. Currently still making a manual step in create methods in services, but working towards pushing it further down the layers of abstraction to avoin the problems associated with accidentally not explicitly asking for the creator to be attached.